move scala server into its own file
